+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+
 develop
  - Started refracturing structure for auditory scales
  -
+
+0.67 -> 0.68
+ - Reading and Writing Audiofiles is now handled using the soundfile library
+ - Fixed some tests and depricated numpy apis
 
 0.66.1 -> 0.67
  - Changed functions in stats to property